backend/requirements.txt
asttokens==2.4.1
backcall==0.2.0
blinker==1.7.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.0
contourpy==1.1.1
cors==1.0.1
cycler==0.12.1
decorator==5.1.1
dill==0.3.7
executing==2.0.1
filelock==3.13.1
Flask==3.0.0
Flask-Cors==4.0.0
fonttools==4.46.0
fsspec==2023.12.2
future==0.18.3
gevent==23.9.1
gitdb==4.0.11
GitPython==3.1.40
greenlet==3.0.2
idna==3.6
importlib-metadata==7.0.0
importlib-resources==6.1.1
ipython==8.12.3
ipywidgets==8.1.1
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
jupyterlab-widgets==3.0.9
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.7.4
matplotlib-inline==0.1.6
mpmath==1.3.0
networkx==3.1
numpy==1.24.4
opencv-python==4.8.1.78
packaging==23.2
pandas==2.0.3
parso==0.8.3
pickleshare==0.7.5
Pillow==10.1.0
prompt-toolkit==3.0.42
psutil==5.9.6
pure-eval==0.2.2
py-cpuinfo==9.0.0
pycparser==2.21
Pygments==2.17.2
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-file==1.5.1
scipy==1.10.1
seaborn==0.13.0
six==1.16.0
smmap==5.0.1
stack-data==0.6.3
sympy==1.12
thop==0.1.1.post2209072238
tldextract==5.1.1
torch==2.1.1
torchvision==0.16.1
tqdm==4.66.1
traitlets==5.14.0
typing_extensions==4.9.0
tzdata==2023.3
ultralytics==8.0.227
urllib3==2.1.0
wcwidth==0.2.12
Werkzeug==3.0.1
widgetsnbextension==4.0.9
zipp==3.17.0
zope.event==5.0
zope.interface==6.1
Gunicorn
